reorganisation, placed Variable within variable.py
